+++ main.py
@@ -1,5 +1,4 @@
 import copy
-import dataclasses
 import datetime
 import pickle
 import typing
@@ -7,48 +6,11 @@ import sys
 import os
 
 from injector import inject
+from models import Feedback
+from infra import Repository
+from infra import Store
 
 DATABASE_FILE = "datafile.pickle"
-
-
-@dataclasses.dataclass(frozen=True)
-class Feedback:
-    feedback_id: int
-    title: str
-    description: str
-    created_at: datetime.datetime
-    updated_at: datetime.datetime
-
-
-@dataclasses.dataclass()
-class Store:
-    feedbacks: typing.List[Feedback] = dataclasses.field(default_factory=list)
-    created_at: datetime.datetime = dataclasses.field(default_factory=datetime.datetime.utcnow)
-
-
-@dataclasses.dataclass()
-class Repository:
-    store: Store = dataclasses.field(default_factory=Store)
-    filename: typing.Optional[str] = None
-
-    def reset(self):
-        self.store = Store()
-
-    @classmethod
-    def load(cls, filename: str) -> "Repository":
-        if os.path.exists(filename):
-            with open(filename, "rb") as file:
-                store = pickle.load(file)
-        else:
-            store = Store()
-        return cls(
-            store=store,
-            filename=filename
-        )
-
-    def persistent(self):
-        with open(self.filename, "wb") as file:
-            pickle.dump(self.store, file)
 
 
 class FeedbackRepository:
